Fixed tests without tensorflow
