+++ tests/tests_agent_parts/estimators/__init__.py
@@ -1,0 +1,3 @@
+# LearnRL a python library to learn and use reinforcement learning
+# Copyright (C) 2020 Mathïs FEDERICO <https://www.gnu.org/licenses/>
+
