+++ tests/tests_agent_parts/estimators/test_estimator.py
@@ -1,3 +1,6 @@
+# LearnRL a python library to learn and use reinforcement learning
+# Copyright (C) 2020 Mathïs FEDERICO <https://www.gnu.org/licenses/>
+
 import pytest
 import sys
 import importlib
@@ -33,23 +36,3 @@ def test_update_learning_rate():
     if evaluation.learning_rate != expected_learning_rate:
         raise ValueError(f"Learning rate is {evaluation.learning_rate} and did not decay to {expected_learning_rate}")
 
-def test_instanciate_keras_estimator_without_tensorflow(hide_tensorflow):
-    observation_space = Discrete(4)
-    action_space = Discrete(3)
-
-    class DummyEstimator(learnrl.estimators.KerasEstimator):
-        def build(self, **kwargs):
-            pass
-
-    with pytest.raises(ImportError, match=r".*tensorflow >= 2.*"):
-        estimator = DummyEstimator(observation_space, action_space)
-
-def test_instanciate_keras_estimator_with_tensorflow():
-    observation_space = Discrete(4)
-    action_space = Discrete(3)
-
-    class DummyEstimator(learnrl.estimators.KerasEstimator):
-        def build(self, **kwargs):
-            pass
-
-    estimator = DummyEstimator(observation_space, action_space)
